cleanup
- putting in SoldierDeadInfoState ...

+++ src/Geoscape/Globe.cpp
@@ -2313,13 +2313,13 @@ void Globe::drawPath(
 	if (fabs(sx)<0.01)
 	{
 //kL		seg = abs( sy/(2*M_PI)*48 );
-		seg = abs( sy / (2.0 * M_PI) * 48.0 ); // kL
+		seg = abs( static_cast<int>(sy / (2.0 * M_PI) * 48.0) ); // kL
 		if (seg == 0) ++seg;
 	}
 	else
 	{
 //kL		seg = abs( sx/(2*M_PI)*96 );
-		seg = abs( sx / (2.0 * M_PI) * 96.0 ); // kL
+		seg = abs( static_cast<int>(sx / (2.0 * M_PI) * 96.0) ); // kL
 		if (seg == 0) ++seg;
 	}
 	sx /= seg;
